% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_metadata.R
\name{gen_meta_obj}
\alias{gen_meta_obj}
\title{Generate metadata as S4 object}
\usage{
gen_meta_obj(archr_meta, sample_meta)
}
\arguments{
\item{archr_meta}{archR readme dataframe.}

\item{sample_meta}{sample metadata sheet.}
}
\value{
S4 object of the two dataframes.
}
\description{
This function converts metadata dataframes to S4 object. It assumes that the format
as generated by gen_sample_sheets() in gmeta package. It requires both archr and sample
metadata dataframes.
}
\examples{
gen_meta_obj(archr_meta, sample_meta)
}
