% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_metadata.R
\name{gen_sample_sheets}
\alias{gen_sample_sheets}
\title{get metadata sample sheets in the right format}
\usage{
gen_sample_sheets()
}
\value{
sample_metadata.txt and archr_metadata.txt files
}
\description{
get metadata sample sheets in the right format
}
\examples{
gen_sample_sheets()
}
